Webapp - final version (correct)/requirements.txt
Flask==2.2.5
Flask_WTF==1.1.1
mysql_connector_repackaged==0.3.1
opencv_contrib_python==4.8.0.74
opencv_python==4.7.0.72
openpyxl==3.1.2
piexif==1.1.3
Pillow==9.3.0
Pillow==10.0.0
pyperclip==1.8.2
pytesseract==0.3.10
python_docx==0.8.11
simplekml==1.3.6
torch==2.0.1+cu118
torchvision==0.15.2+cu118
Werkzeug==2.3.6
WTForms==3.0.1
XlsxWriter==3.1.2
